* commit '25da2aba928db60226574139a3aaa678c76938ec': (90 commits)
  small login fix
  - Bug fixed - boxplot diagrams login-page
  error messages for login page
  Update finished users fixed
  fix typo
  device fix trash can
  pretty login page
  Graph from genuine result data - only for tested devices
  Generating graphs using the device names, gotten from through TestObjectService -> homeController model attribute -> thymeleaf inline javascript -> highcharts javascript chart generation
  Passing objects using the model attributes
  - Fitts Test included and working with random values
  sprint3 _ Security working
  databaseloader, testplanController + messages adapted
  survey version 2 done
  fixing merges
  fixing merges
  fixing merges
  device fix (merge)
  device fix
  -Aanmaken van dud testresultaten wanneer sequences worden 'uitgevoerd' (tijdelijke implementatie ter vervanging van fitts law test -Updaten van TestPlan bij het vervolledigen van test
  ...

+++ src/main/java/be/uantwerpen/fti/se/service/UserService.java
@@ -1,9 +1,14 @@
 package be.uantwerpen.fti.se.service;
 
+import be.uantwerpen.fti.se.model.TestPlan;
 import be.uantwerpen.fti.se.model.User;
 import be.uantwerpen.fti.se.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
 
 /**
  * Created by Edwin on 6/10/2015.
@@ -17,12 +22,30 @@ public class UserService {
         return this.userRepository.findAll();
     }
 
+    public List<User> findAllAdmins() {
+
+        List<User> userList = new ArrayList<User>();
+        for (User  user : userRepository.findAll()) {
+            if(user.isAdmin())
+                userList.add(user);
+        }
+        return userList;
+    }
+
     public void add(final User user) {
         this.userRepository.save(user);
     }
 
     public void delete(Long id) {
-        this.userRepository.delete(id);
+        if(findOne(id).isAdmin()) {
+            int admins = findAllAdmins().size();
+            if (((List<?>) findAllAdmins()).size() > 1) {
+                this.userRepository.delete(id);
+            }
+        }
+        else{
+            this.userRepository.delete(id);
+        }
     }
